+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `LocationCountry` label is now displayed.
+- Failure message will no longer show up when starting to type on `LocationSearch`.
+
+### Added
+- Label id prop for `LocationCountry`.
+- Loading state after selecting an address on `LocationSearch`.
 
 ## [0.13.0] - 2020-11-17
 ### Added